requirements.txt
crewai>=0.16.2
langchain>=0.1.0
openai>=1.6.1
python-dotenv>=1.0.0
pyyaml>=6.0.1 
streamlit>=1.25.0